Add texts to preview
